case_studies/cs5_whale_reid/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
pillow>=9.0.0
opencv-python>=4.5.0
tqdm>=4.62.0
requests>=2.26.0

# Model dependencies
transformers>=4.30.0
timm>=0.9.0
huggingface-hub>=0.14.0

# Wildlife datasets
wildlife-datasets>=0.1.0

# Image processing and visualization
matplotlib>=3.4.0
scikit-learn>=1.0.0
scipy>=1.7.0

# For Grounding DINO (install separately)
# groundingdino - see installation instructions in README

# Additional utilities
pyyaml>=5.4.0
